clean up index.html, delete unused favicons, move Recording.js to comps folder
